+++ apphelpers/rest/hug.py
@@ -1,14 +1,18 @@
-import inspect
 from dataclasses import dataclass, asdict
-from functools import wraps
 from falcon import HTTPUnauthorized, HTTPForbidden, HTTPNotFound
-
 import hug
-import hug.introspect as introspect
+from hug.decorators import wraps
 
 from apphelpers.db.peewee import dbtransaction
 from apphelpers.errors import InvalidSessionError
 from apphelpers.sessions import SessionDBHandler
+from converge import settings
+
+from apphelpers.loggers import api_logger
+
+if settings.get('HONEYBADGER_API_KEY'):
+    from honeybadger import Honeybadger
+    from honeybadger.utils import filter_dict
 
 
 def phony(f):
@@ -27,6 +31,33 @@ def raise_not_found_on_none(f):
     return f
 
 
+def honeybadger_wrapper(hb):
+    """
+    wrapper that executes the function in a try/except
+    If an exception occurs, it is first reported to Honeybadger
+    """
+    def wrapper(f):
+        @wraps(f)
+        def f_wrapped(*args, **kw):
+            try:
+                ret = f(*args, **kw)
+            except Exception as e:
+                try:
+                    hb.notify(
+                        e,
+                        context={
+                            'func': f.__name__,
+                            'args': args,
+                            'kwargs': filter_dict(kw, settings.HB_PARAM_FILTERS)
+                        }
+                    )
+                finally:
+                    raise e
+            return ret
+        return f_wrapped
+    return wrapper
+
+
 @hug.directive()
 def user_id(default=None, request=None, **kwargs):
     return request.context['user'].id
@@ -42,6 +73,16 @@ def user_email(default=None, request=None, **kwargs):
     return request.context['user'].email
 
 
+@hug.directive()
+def domain(default=None, request=None, **kwargs):
+    return request.headers['HOST']
+
+
+@hug.directive()
+def user_mobile(default=None, request=None, **kwargs):
+    return request.context['user'].mobile
+
+
 @dataclass
 class User:
     sid: str=None
@@ -49,6 +90,7 @@ class User:
     name: str=None
     groups: tuple=()
     email: str=None
+    mobile: str=None
     site_groups: dict=None
 
     def to_dict(self):
@@ -62,20 +104,24 @@ def setup_strict_context_setter(sessions):
 
     def set_context(token):
 
-        uid, groups, name, email, site_groups = None, [], '', None, {}
+        uid, groups, name, email, mobile, site_groups = None, [], '', None, None, {}
 
         if token:
             try:
-                session = sessions.get(token, ['uid', 'name', 'groups', 'email', 'site_groups'])
-                uid, name, groups, email, site_groups = (
+                session = sessions.get(
+                    token,
+                    ['uid', 'name', 'groups', 'email', 'mobile', 'site_groups']
+                )
+                uid, name, groups, email, mobile, site_groups = (
                     session['uid'], session['name'], session['groups'],
-                    session['email'], session['site_groups']
+                    session['email'], session['mobile'], session['site_groups']
                 )
             except InvalidSessionError:
                 raise HTTPUnauthorized('Invalid or expired session')
 
         return User(
-            sid=token, id=uid, name=name, groups=groups, email=email, site_groups=site_groups
+            sid=token, id=uid, name=name, groups=groups,
+            email=email, mobile=mobile, site_groups=site_groups
         )
 
     return set_context
@@ -88,20 +134,28 @@ def setup_context_setter(sessions):
         Only sets context based on session.
         Does not raise any error
         """
-        uid, groups, name, email, site_groups = None, [], '', None, {}
+        uid, groups, name, email, mobile, site_groups = None, [], '', None, None, {}
         token = request.get_header('Authorization')
         if token:
             try:
-                session = sessions.get(token, ['uid', 'name', 'groups', 'email', 'site_groups'])
-                uid, name, groups, email, site_groups = (
-                    session['uid'], session['name'], session['groups'],
-                    session['email'], session['site_groups']
+                session = sessions.get(
+                    token,
+                    ['uid', 'name', 'groups', 'email', 'mobile', 'site_groups']
                 )
+                uid, name, groups, email, mobile, site_groups = (
+                    session['uid'], session['name'], session['groups'],
+                    session['email'], session['mobile'], session['site_groups']
+                )
+                if api_logger:
+                    api_logger.info('{} | {} | {}',
+                                    uid, request.method, request.url)
+
             except InvalidSessionError:
                 pass
 
         request.context['user'] = User(
-            sid=token, id=uid, name=name, groups=groups, email=email, site_groups=site_groups
+            sid=token, id=uid, name=name, groups=groups,
+            email=email, mobile=mobile, site_groups=site_groups
         )
     return set_context
 
@@ -116,6 +170,7 @@ class APIFactory:
         self.multi_site_enabled = False
         self.site_identifier = None
         self.urls_prefix = urls_prefix
+        self.honeybadger_wrapper = phony
 
     def enable_multi_site(self, site_identifier):
         self.multi_site_enabled = True
@@ -123,6 +178,17 @@ class APIFactory:
 
     def setup_db_transaction(self, db):
         self.db_tr_wrapper = dbtransaction(db)
+
+    def setup_honeybadger_monitoring(self):
+        api_key = settings.HONEYBADGER_API_KEY
+        if not api_key:
+            print("Info: Honeybadger API KEY not found. Honeybadger not set")
+            return
+
+        print("Info: Setting up Honeybadger")
+        hb = Honeybadger()
+        hb.configure(api_key=api_key)
+        self.honeybadger_wrapper = honeybadger_wrapper(hb)
 
     def setup_session_db(self, sessiondb_conn):
         """
@@ -142,8 +208,9 @@ class APIFactory:
             login_required = getattr(f, 'login_required', None)
             groups_required = getattr(f, 'groups_required', None)
             groups_forbidden = getattr(f, 'groups_forbidden', None)
+            authorizer = getattr(f, 'authorizer', None)
 
-            if login_required or groups_required or groups_forbidden:
+            if login_required or groups_required or groups_forbidden or authorizer:
 
                 @wraps(f)
                 def wrapper(request, *args, **kw):
@@ -163,6 +230,9 @@ class APIFactory:
                     if groups_forbidden and groups.intersection(groups_forbidden):
                         raise HTTPForbidden('Unauthorized access')
 
+                    if authorizer and not authorizer(user, *args, **kw):
+                        raise HTTPForbidden('Unauthorized access')
+
                     return f(*args, **kw)
             else:
                 wrapper = f
@@ -176,8 +246,9 @@ class APIFactory:
             login_required = getattr(f, 'login_required', None)
             groups_required = getattr(f, 'groups_required', None)
             groups_forbidden = getattr(f, 'groups_forbidden', None)
+            authorizer = getattr(f, 'authorizer', None)
 
-            if login_required or groups_required or groups_forbidden:
+            if login_required or groups_required or groups_forbidden or authorizer:
 
                 @wraps(f)
                 def wrapper(request, *args, **kw):
@@ -200,6 +271,9 @@ class APIFactory:
                     if groups_forbidden and groups.intersection(groups_forbidden):
                         raise HTTPForbidden('Unauthorized access')
 
+                    if authorizer and not authorizer(user, *args, **kw):
+                        raise HTTPForbidden('Unauthorized access')
+
                     return f(*args, **kw)
 
             else:
@@ -216,7 +290,10 @@ class APIFactory:
     def build(self, method, method_args, method_kw, f):
         print(f'{method_args[0]} [{method.__name__.upper()}] => {f.__module__}:{f.__name__}')
         m = method(*method_args, **method_kw)
-        f = self.access_wrapper(self.db_tr_wrapper(raise_not_found_on_none(f)))
+        f = self.access_wrapper(
+            self.honeybadger_wrapper(
+                self.db_tr_wrapper(
+                    raise_not_found_on_none(f))))
         # NOTE: ^ wrapper ordering is important. access_wrapper needs request which
         # others don't. If access_wrapper comes late in the order it won't be passed
         # request parameter.