+++ tests/app/endpoints.py
@@ -53,6 +53,24 @@ echo_multisite_groups.groups_required = [sitegroups.privileged.value]
 echo_multisite_groups.groups_forbidden = [sitegroups.forbidden.value]
 
 
+def process_request(request, body):
+    return {'body': body, 'headers': request.headers}
+
+
+def process_raw_request(request):
+    return {'raw_body': request.stream.read().decode(), 'headers': request.headers}
+process_raw_request.not_found_on_none = True
+
+
+def check_authorization(user, *args, **kw):
+    return kw['word'] == 'authorized'
+
+
+def custom_authorization_echo(word):
+    return word
+custom_authorization_echo.authorizer = check_authorization
+
+
 def setup_routes(factory):
 
     factory.get('/echo/{word}')(echo)
@@ -61,6 +79,7 @@ def setup_routes(factory):
     factory.get('/add')(add)
 
     factory.get('/secure-echo/{word}')(secure_echo)
+    factory.get('/custom-authorization-echo/{word}')(custom_authorization_echo)
     factory.get('/echo-groups')(echo_groups)
 
     factory.post('/me/uid')(get_my_uid)
@@ -70,6 +89,9 @@ def setup_routes(factory):
     factory.get('/sites/{site_id}/secure-echo/{word}')(secure_multisite_echo)
     factory.get('/sites/{site_id}/echo-groups')(echo_multisite_groups)
     factory.get('/sites/{site_id}/snakes/{name}')(get_secure_snake)
+
+    factory.post('/request-and-body')(process_request)
+    factory.post('/request-raw-body', parse_body=False)(process_raw_request)
 
     # ar_handlers = (None, arlib.create, None, arlib.get, arlib.update, None)
     # factory.map_resource('/resttest/', handlers=ar_handlers)