+++ CHANGELOG.rst
@@ -2,6 +2,23 @@
 History
 =======
 
+0.21.1 (2022-06-09)
+-------------------
+* Fix for SMTP+SSL connection
+
+0.21.0 (2022-05-18)
+-------------------
+* applogger: general purpose application logging
+
+0.20.0 (2022-04-29)
+-------------------
+* Support for custom authorizaion
+
+0.19.1 (2021-10-07)
+-------------------
+
+* Report function args in honeybadger context
+
 0.9.2 (2019-05-20)
 ------------------